Even more refactoring, but hopefully that's all for 1.20.4?

+++ api/src/main/java/mods/thecomputerizer/theimpossiblelibrary/api/world/PosHelper.java
@@ -10,7 +10,7 @@ public class PosHelper {
     }
 
     public static BlockPosAPI<?> getPos(double x, double y, double z) {
-        return getPos(new Vector3((int)x,(int)y,(int)z));
+        return getPos(new Vector3(x,y,z));
     }
 
     public static BlockPosAPI<?> getPos(int x, int y, int z) {
